# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IoT 物联网平台后端 API（Go 重构版）",
        "title": "IoT Platform Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@iot-platform.local.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.7.0"
    },
    "host": "localhost:8182",
    "basePath": "/",
    "paths": {
...
        "/api/devices/{deviceId}/heartbeat": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "设备心跳",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "DeviceAuth": []
                    }
                ]
            }
        },
...
        "/api/devices/{deviceId}/sensorData": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "查询设备传感器数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path"
                    },
                    {
                        "type": "integer",
                        "description": "数量限制，默认50",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "起始时间（RFC3339），默认3天前",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间（RFC3339），默认当前时间",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "上报传感器数据",
                "parameters": [
                    {
                        "description": "传感器数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DeviceStatusDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    }
                },
                "security": [
                    {
                        "DeviceAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "common.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "entity.DeviceStatusDTO": {
            "type": "object",
            "required": [
                "sensors"
            ],
            "properties": {
                "sensors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.SensorData"
                    }
                }
            }
        },
...
        "entity.SensorData": {
            "type": "object",
            "required": [
                "name",
                "type",
                "value"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
    },
    "securityDefinitions": {
        "DeviceAuth": {
            "description": "设备 Token（设备注册后获取）",
            "type": "apiKey",
            "name": "X-Device-Token",
            "in": "header"
        },
        "UserAuth": {
            "description": "Bearer Token（用户 Sa-Token），登录后获取",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}